requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-Cors==4.0.1
python-socketio==5.11.4
eventlet==0.36.1
python-dotenv==1.0.1
anthropic==0.39.0
gunicorn==22.0.0
PyJWT==2.9.0
